apps/bot.metrics/requirements.txt
apache-airflow-providers-papermill==2.2.0
ipykernel==5.5.6
psycopg2-binary==2.9.3